refactor: rename sample view files → debug_item_*, class SampleItemListView → DebugItemListView
