+++ doc/plans/restyle-debug-views.md
@@ -9,9 +9,9 @@ shadcn_ui (the project's design system), and rename the directory to `debug_feat
 
 | File | Material widgets used | Already shadcn |
 |------|-----------------------|----------------|
-| `sample_item_list_view.dart` (~130 lines) | `Scaffold`, `AppBar`, `ListTile`, `IconButton` | `ShadButton`, lucide icons mixed in |
-| `sample_item_details_view.dart` (~310 lines) | `Scaffold`, `AppBar`, `OutlinedButton`, `LinearProgressIndicator`, `Theme.of(...).textTheme.titleMedium` | `ShadButton`, `ShadDialog`, `ShadInput` |
-| `scale_debug_view.dart` (~140 lines) | `Scaffold`, `AppBar`, `FilledButton` (×7), `Theme.of(...).textTheme.titleMedium` | `ShadButton` (one instance) |
+| `debug_item_list_view.dart` (~130→~280 lines) | `Scaffold`, `AppBar`, `ListTile`, `IconButton` | `ShadButton`, lucide icons mixed in |
+| `debug_item_details_view.dart` (~310→~430 lines) | `Scaffold`, `AppBar`, `OutlinedButton`, `LinearProgressIndicator`, `Theme.of(...).textTheme.titleMedium` | `ShadButton`, `ShadDialog`, `ShadInput` |
+| `scale_debug_view.dart` (~140→~170 lines) | `Scaffold`, `AppBar`, `FilledButton` (×7), `Theme.of(...).textTheme.titleMedium` | `ShadButton` (one instance) |
 
 None of the files have tests beyond `sample_item_details_view_test.dart` (exists but may be minimal).
 
@@ -19,9 +19,9 @@ None of the files have tests beyond `sample_item_details_view_test.dart` (exists
 
 | File | Import |
 |------|--------|
-| `lib/src/app.dart` | `sample_item_list_view.dart`, `sample_item_details_view.dart`, `scale_debug_view.dart` |
-| `lib/src/settings/advanced_page.dart` | `sample_item_list_view.dart` |
-| `test/sample_feature/sample_item_details_view_test.dart` | `sample_item_details_view.dart` |
+| `lib/src/app.dart` | `debug_item_list_view.dart`, `debug_item_details_view.dart`, `scale_debug_view.dart` |
+| `lib/src/settings/advanced_page.dart` | `debug_item_list_view.dart` |
+| `test/debug_feature/debug_item_details_view_test.dart` | `debug_item_details_view.dart` |
 
 ## Tasks
 
@@ -29,34 +29,55 @@ None of the files have tests beyond `sample_item_details_view_test.dart` (exists
 
 - `git mv lib/src/sample_feature lib/src/debug_feature`
 - `git mv test/sample_feature test/debug_feature`
-- Update imports in `lib/src/app.dart`, `lib/src/settings/advanced_page.dart`, `test/debug_feature/sample_item_details_view_test.dart`
+- Update imports in `lib/src/app.dart`, `lib/src/settings/advanced_page.dart`, `test/debug_feature/debug_item_details_view_test.dart`
 - Update internal imports within the 3 view files
-- Run `flutter test` + `flutter analyze` — should be a no-op rename, 0 failures expected
+- Rename `SampleItemListView` class → `DebugItemListView`
+- Rename files: `sample_item_list_view.dart` → `debug_item_list_view.dart`, `sample_item_details_view.dart` → `debug_item_details_view.dart`
+- Run `flutter test` + `flutter analyze`
 
 ### Task 2: Restyle `debug_item_list_view.dart`
 
-Replace Material → shadcn:
-- `Scaffold` → `ShadApp` layout or inline (this is a pushed route, already inside a `ShadApp`)
-- `AppBar` → shadcn app-bar equivalent or a `ShadAppBar`
-- `ListTile` → `ShadCard` or `ShadListTile` pattern
-- `IconButton` → `ShadButton.ghost`
+**Direction**: Ambitious — improve UX while keeping a dense list layout.
+
+Replace Material → shadcn with UX enhancements:
+
+- **Custom header row**: No `Scaffold`/`AppBar`. A `Row` with title text and a scan button. When idle: radar icon. When scanning: spinner icon + "Scanning..." text.
+- **Scanning indicator**: `ShadProgress` bar at top while scanning, driven by `DeviceController.scanningStream`.
+- **Empty state**: When 0 devices and not scanning — centered "No devices discovered — tap Scan to search" message.
+- **Device grouping**: Section headers for Machines, Scales, Sensors. Only show non-empty groups.
+- **Dense list items**: Custom `Row`-based item (no `ListTile`) with connection-state icon, device name/id, and `[Inspect]` `[Connect]` actions.
+- **Sensors**: Display in list but no Connect button (sensors don't have a connect flow).
+- **Device count**: Show count in header ("Debug · 3 devices").
 
 ### Task 3: Restyle `debug_item_details_view.dart`
 
-Replace Material → shadcn:
-- `AppBar` → shadcn equivalent
-- `OutlinedButton` (Wake/Sleep) → `ShadButton.outline`
-- `LinearProgressIndicator` → `ShadProgress`
-- `Theme.of(context).textTheme.titleMedium` → shadcn typography (`Text` with `style: ShadTheme.of(context).textTheme.h4` or similar)
-- Section headers (`"Shot snapshot:"`, `"Shot settings:"`, etc.) → use `ShadTextTheme` or `Text` with bold style
+**Direction**: Ambitious — responsive layout, card-based sections, state dropdown.
+
+Replace Material → shadcn with UX enhancements:
+
+- **Responsive layout**: Horizontal `Flex` on wide screens (tablet), stacked vertically on narrow (phone). Use `LayoutBuilder`.
+- **Section cards**: Each data section (shot snapshot, shot settings, water levels, machine info) wrapped in `ShadCard` with titled header replacing bare `Theme.of(context).textTheme.titleMedium`.
+- **Machine state control**: Replace `OutlinedButton` (Wake/Sleep) with a `ShadSelect<MachineState>` dropdown of `MachineState.values`, calling `requestState()` on change. Also show current state as text.
+- **`LinearProgressIndicator`** → `ShadProgress`.
+- **Typography**: Replace `Theme.of(context).textTheme.titleMedium` with shadcn text styles.
+- **Firmware update**: Keep existing flow (already uses `ShadDialog`), just restyle the trigger button.
+- **Serial comms**: Leave as-is for now.
 
 ### Task 4: Restyle `scale_debug_view.dart`
 
-Replace Material → shadcn:
-- `AppBar` → shadcn equivalent
-- `FilledButton` (×7) → `ShadButton` (primary variant), or `ShadButton.outline` for secondary actions
-- `Theme.of(context).textTheme.titleMedium` → shadcn typography
-- The lone inline `ShadButton` with missing trailing comma — fix formatting
+**Direction**: Ambitious — weight hero, button hierarchy, custom header.
+
+Replace Material → shadcn with UX enhancements:
+
+- **Custom header**: Replace `Scaffold`/`AppBar` with a header row showing scale name + device ID.
+- **Weight hero**: Large centered `Text` for current weight (e.g. "247.3g"). Prominent display since it's the primary debug data point.
+- **Button hierarchy**: 
+  - **Primary actions** (top): Tare (`ShadButton`) + Disconnect (`ShadButton.destructive`)
+  - **Display controls** (row): Wake (`ShadButton.outline`) · Sleep (`ShadButton.outline`)
+  - **Timer controls** (row): Start (`ShadButton.outline`) · Stop (`ShadButton.outline`) · Reset (`ShadButton.outline`)
+- **Info row**: Battery level (plain text) + update latency (plain text) below weight.
+- **Typography**: Replace `Theme.of(context).textTheme.titleMedium` with shadcn text styles.
+- **States**: Show "Connecting…" while waiting; show weight data when active; show "Waiting for data…" otherwise.
 
 ### Task 5: Final verification
