+++ lib/src/debug_feature/debug_item_details_view.dart
